Moved storage implementations to dedicated files
